+++ src/ivr/handler/accountInfo/account2.js
@@ -1,10 +1,10 @@
-const productsList = require("../options/products");
-const offersList = require("../options/offers");
-const accountInfo1 = require("../options/accountInfo1");
-const accountInfo2 = require("../options/accountInfo2");
-const redirectWelcome = require("../options/redirectWelcome")
-const accountMenu = require("../options/accountMenu")
-const customerRep = require("../options/customerRep")
+const productsList = require("../../options/products");
+const offersList = require("../../options/offers");
+const accountInfo1 = require("../../options/accountInfo1");
+const accountInfo2 = require("../../options/accountInfo2");
+const redirectWelcome = require("../../options/redirectWelcome")
+const accountMenu = require("../../options/accountMenu")
+const customerRep = require("../../options/customerRep")
 
 module.exports = function account2(digits) {
     console.log("Here")
